testing damagable

+++ Assets/FangZheng/Scripts/Weapons/Weapon.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 public abstract class Weapon : MonoBehaviour
 {
     [SerializeField] protected WeaponData data;
-    protected float damage;
+    protected int damage;
     protected LayerMask hitLayers;
 
     protected virtual void Awake()
